+++ app/src/main/java/com/example/mytalkcharge/repository/WeatherResponse.kt
@@ -50,7 +50,7 @@ data class Main(
     val feels_like: Double,
     val grnd_level: Int,
     val humidity: Int,
-    val pressure: Int,
+    val pressure: Float,
     val sea_level: Int,
     val temp: Double,
     val temp_kf: Double,
